// frontend/src/api/model/productDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CoordinatesDto } from './coordinatesDto';
import type { ProductDtoUnitOfMeasure } from './productDtoUnitOfMeasure';
import type { OrganizationDto } from './organizationDto';
import type { PersonDto } from './personDto';

/**
 * объект
 */
export interface ProductDto {
  id: number;
  name?: string;
  coordinates?: CoordinatesDto;
  creationDate?: string;
  unitOfMeasure?: ProductDtoUnitOfMeasure;
  manufacturer?: OrganizationDto;
  price?: number;
  manufactureCost?: number;
  rating?: number;
  partNumber?: string;
  owner?: PersonDto;
}