// frontend/src/api/model/productDtoUnitOfMeasure.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ProductDtoUnitOfMeasure = typeof ProductDtoUnitOfMeasure[keyof typeof ProductDtoUnitOfMeasure];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProductDtoUnitOfMeasure = {
  METERS: 'METERS',
  CENTIMETERS: 'CENTIMETERS',
  LITERS: 'LITERS',
  GRAMS: 'GRAMS',
  MILLIGRAMS: 'MILLIGRAMS',
} as const;